+++ srcjs/mouse.js
@@ -1,6 +1,13 @@
 export default class MouseHandler {
-  constructor(svgid, tooltipHandler, hoverHandlers, selectionHandlers) {
+  constructor(
+    svgid,
+    nearestHandler,
+    tooltipHandler,
+    hoverHandlers,
+    selectionHandlers
+  ) {
     this.svgid = svgid;
+    this.nearestHandler = nearestHandler;
     this.tooltipHandler = tooltipHandler;
     this.hoverHandlers = hoverHandlers;
     this.selectionHandlers = selectionHandlers;
@@ -26,8 +33,10 @@ export default class MouseHandler {
     // add listeners
     const svgNode = document.getElementById(this.svgid);
     svgNode.addEventListener('mouseover', this, true);
-    svgNode.addEventListener('mousemove', this, true);
+    svgNode.addEventListener('mousemove', this);
     svgNode.addEventListener('mouseout', this, true);
+    svgNode.addEventListener('mousedown', this, true);
+    svgNode.addEventListener('wheel', this, true);
     svgNode.addEventListener('click', this, true);
 
     // return true to add to list of handLers
@@ -41,6 +50,8 @@ export default class MouseHandler {
       svgNode.removeEventListener('mouseover', this);
       svgNode.removeEventListener('mousemove', this);
       svgNode.removeEventListener('mouseout', this);
+      svgNode.removeEventListener('mousedown', this);
+      svgNode.removeEventListener('wheel', this);
       svgNode.removeEventListener('click', this);
     } catch (e) {
       console.error(e);
@@ -50,29 +61,43 @@ export default class MouseHandler {
   handleEvent(event) {
     try {
       const target = event.target;
-      const mousePos = new DOMPoint(event.pageX, event.pageY);
+      let handled = false,
+        nearest = null;
       if (event.type === 'mouseout') {
-        if (this.svgid !== target.id) {
-          this.mouseOnHandlers.forEach(function (h) {
-            h.clear();
-          });
-        }
-      } else if (event.type === 'mouseover') {
-        if (this.svgid !== target.id) {
-          this.mouseOnHandlers.forEach(function (h) {
-            h.applyOn(target, mousePos);
-          });
-        }
-      } else if (event.type === 'mousemove') {
+        this.mouseOnHandlers.forEach(function (h) {
+          h.clear();
+        });
+      } else if (event.type === 'mouseover' && !event.buttons) {
+        this.mouseOnHandlers.forEach(function (h) {
+          h.applyOn(target, event);
+        });
+      } else if (event.type === 'mousemove' && !event.buttons) {
         if (this.svgid !== target.id) {
           if (this.tooltipHandler) {
-            this.tooltipHandler.applyOn(target, mousePos);
+            handled = this.tooltipHandler.applyOn(target, event);
           }
+        }
+        if (this.nearestHandler && !handled) {
+          event.fromNearest = true;
+          nearest = this.nearestHandler.applyOn(target, event);
+          if (nearest) {
+            this.mouseOnHandlers.forEach(function (h) {
+              h.applyOn(nearest, event);
+            });
+          } else {
+            this.mouseOnHandlers.forEach(function (h) {
+              h.clear();
+            });
+          }
+        }
+      } else if (event.type === 'mousedown' || event.type === 'wheel') {
+        if (this.tooltipHandler) {
+          this.tooltipHandler.clear(event);
         }
       } else if (event.type === 'click') {
         if (this.svgid !== target.id) {
           this.selectionHandlers.forEach(function (h) {
-            h.applyOn(target);
+            h.applyOn(target, event);
           });
         }
       }