+++ srcjs/tooltip.js
@@ -1,5 +1,9 @@
 import * as d3 from 'd3';
-import { getWindowViewport, getHTMLElementMatrix } from './geom';
+import {
+  getWindowViewport,
+  getHTMLElementMatrix,
+  rectContainsRect
+} from './geom';
 
 export default class TooltipHandler {
   constructor(
@@ -79,17 +83,21 @@ export default class TooltipHandler {
   }
 
   isValidTarget(target) {
-    return target instanceof SVGGraphicsElement && target.hasAttribute('title');
+    return (
+      target instanceof SVGGraphicsElement &&
+      !(target instanceof SVGSVGElement) &&
+      target.hasAttribute('title')
+    );
   }
 
-  applyOn(target, mousePos) {
+  applyOn(target, event) {
     try {
       if (this.isValidTarget(target)) {
-        const svgNode = target.ownerSVGElement;
+        const mousePos = new DOMPoint(event.pageX, event.pageY);
         const tooltipEl = d3.select('div.' + this.clsName);
         let tooltipPos;
         if (target.id === this.lastTargetId) {
-          tooltipPos = this.calculatePosition(tooltipEl, svgNode, mousePos);
+          tooltipPos = this.calculatePosition(tooltipEl, target, mousePos);
           tooltipEl
             .style('left', tooltipPos.x + 'px')
             .style('top', tooltipPos.y + 'px');
@@ -105,7 +113,7 @@ export default class TooltipHandler {
           }
           tooltipEl.html(this.decodeContent(target.getAttribute('title')));
 
-          tooltipPos = this.calculatePosition(tooltipEl, svgNode, mousePos);
+          tooltipPos = this.calculatePosition(tooltipEl, target, mousePos);
           tooltipEl
             .style('left', tooltipPos.x + 'px')
             .style('top', tooltipPos.y + 'px')
@@ -127,8 +135,9 @@ export default class TooltipHandler {
     return this.decodingTextarea.value;
   }
 
-  calculatePosition(tooltipEl, svgNode, mousePos) {
+  calculatePosition(tooltipEl, target, mousePos) {
     let p, matrix;
+    const svgNode = target.ownerSVGElement;
     const tooltipNode = tooltipEl.node();
     const containerNode = svgNode.parentNode;
     if (this.usecursor) {
@@ -157,39 +166,66 @@ export default class TooltipHandler {
         maxp = maxp.matrixTransform(matrix);
       }
 
-      // calculate horizontal position
+      // target node bounding rect
+      const brect = target.getBoundingClientRect();
+
+      // calculate possible horizontal positions
+      const possibleX = [];
       const spaceRight = maxp.x - (p.x + this.offx);
       const spaceLeft = p.x - this.offx - minp.x;
       if (spaceRight >= tooltipWidth) {
         // fits on right
-        p.x = Math.max(minp.x, p.x + this.offx);
-      } else if (spaceLeft >= tooltipWidth) {
+        possibleX.push(Math.max(minp.x, p.x + this.offx));
+      }
+      if (spaceLeft >= tooltipWidth) {
         // fits on left
-        p.x = Math.min(maxp.x - tooltipWidth, p.x - this.offx - tooltipWidth);
-      } else {
-        // set at middle
-        p.x = Math.max(
-          minp.x,
-          Math.min(maxp.x, p.x + tooltipWidth / 2) - tooltipWidth
+        possibleX.push(
+          Math.min(maxp.x - tooltipWidth, p.x - this.offx - tooltipWidth)
         );
       }
+      // set at middle, as fallback
+      possibleX.push(
+        Math.max(
+          minp.x,
+          Math.min(maxp.x, p.x + tooltipWidth / 2) - tooltipWidth
+        )
+      );
 
-      // calculate vertical position
+      // calculate possible vertical positions
+      const possibleY = [];
       const spaceBottom = maxp.y - (p.y + this.offy);
       const spaceTop = p.y - this.offy - minp.y;
       if (spaceBottom >= tooltipHeight) {
         // fits on bottom
-        p.y = Math.max(minp.y, p.y + this.offy);
-      } else if (spaceTop >= tooltipHeight) {
+        possibleY.push(Math.max(minp.y, p.y + this.offy));
+      }
+      if (spaceTop >= tooltipHeight) {
         // fits on top
-        p.y = Math.min(maxp.y - tooltipHeight, p.y - this.offy - tooltipHeight);
-      } else {
-        // set at middle
-        p.y = Math.max(
-          minp.y,
-          Math.min(maxp.y, p.y + tooltipHeight / 2) - tooltipHeight
+        possibleY.push(
+          Math.min(maxp.y - tooltipHeight, p.y - this.offy - tooltipHeight)
         );
       }
+      // set at middle, as fallback
+      possibleY.push(
+        Math.max(
+          minp.y,
+          Math.min(maxp.y, p.y + tooltipHeight / 2) - tooltipHeight
+        )
+      );
+
+      // combine them into possible tooltip rects
+      const possibleRects = [];
+      possibleX.forEach(function (x) {
+        possibleY.forEach(function (y) {
+          possibleRects.push(new DOMRect(x, y, tooltipWidth, tooltipHeight));
+        });
+      });
+
+      // find the first rect that does not contain the target's rect
+      let found = possibleRects.find((r) => !rectContainsRect(r, brect));
+      if (!found) found = possibleRects[0];
+      p.x = found.x;
+      p.y = found.y;
     } else {
       // Fixed position
       p = new DOMPoint(this.offx, this.offy);